2.0.4 (#38)
* add table mapping keys to threads

* clarifications

* try more padding

* move sqlite3 finalize

* comment typo

* remove unused errno

* bump version number

* remove dup log entry

* add lzip to compprogs

* fix llvm/clang warnings

* uninitialized var

* fix error message

* ti_dirlimit should be off_t

* size_t for read/write

* errno is a macro

* clean up clang warnings

* clean up clang warnings

* longer usleep

* malloc ti_task space

* missed one

* retmax is unused in dfs thread

* add dfs/retmin recommendation notice

* lower exp scan rate

* add detail to warning message

+++ sentinalpipe.c
@@ -174,7 +174,7 @@ static void systemd_signals(void)
 
 	sacatch.sa_handler = sigcatch;
 	sigemptyset(&sacatch.sa_mask);
-	sacatch.sa_flags = SA_RESETHAND;
+	sacatch.sa_flags = (int)SA_RESETHAND;
 
 	for(i = 1; i < NSIG; i++)
 		sigaction(i, &sacatch, NULL);
