+++ signals.c
@@ -35,11 +35,11 @@ void parentsignals(void)
 
 	saparent.sa_handler = sigparent;
 	sigemptyset(&saparent.sa_mask);
-	saparent.sa_flags = SA_RESETHAND;
+	saparent.sa_flags = (int)SA_RESETHAND;
 
 	sareject.sa_handler = sigreject;
 	sigemptyset(&sareject.sa_mask);
-	sareject.sa_flags = SA_RESETHAND;
+	sareject.sa_flags = (int)SA_RESETHAND;
 
 	sigaction(SIGSEGV, NULL, &sasigsegv);
 	sigaction(SIGTSTP, NULL, &sasigstp);
